chore: 빈라인 추가 및 유효성 분리, 예외 추가

+++ src/main/java/roomescape/reservationtime/ReservationTimeService.java
@@ -3,15 +3,20 @@ package roomescape.reservationtime;
 
 import org.springframework.stereotype.Service;
 import roomescape.entities.ReservationTime;
+import roomescape.errors.ErrorCode;
+import roomescape.exceptions.SpringRoomException;
+import roomescape.repositories.ReservationRepository;
 import roomescape.repositories.ReservationTimeRepository;
 
 import java.util.List;
 
 @Service
 public class ReservationTimeService {
+    private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
@@ -24,6 +29,9 @@ public class ReservationTimeService {
     }
 
     public void cancelReservationTime(Long id){
+        if (reservationRepository.findByReservationTimeId(id) != null) {
+            throw new SpringRoomException(ErrorCode.RESERVATION_TIME_CANNOT_BE_DELETED);
+        }
         reservationTimeRepository.deleteById(id);
     }
